refactor(coderd/x/chatd): fold context resource listing into context_prompt.go
Consolidate the chat context resource listing (ContextResources,
pinnedContextResources, contextResourceKind) into context_prompt.go next to
the prompt builder, removing context_detail.go. The two paths walk the same
pinned rows and now share the body decode-and-validate helpers
(decodeInstructionContent, decodeSkillIdentity), so an instruction file or
skill is interpreted identically whether it feeds the prompt or the API
resource list. Tests move alongside the code; no behavior change.
